backend/requirements.txt
fastapi
uvicorn
python-multipart
transformers
flask
torch
pytesseract
pdf2image
tabula-py
opencv-python-headless
numpy
pandas
PyMuPDF
PyPDF2
Pillow
protobuf
sacremoses
rake-nltk
nltk
pdfplumber==0.7.4

